Rename 'exportComponents' to 'components' and move the code into 'components.js'
