+++ packages/cli/src/commands/components.js
@@ -4,7 +4,7 @@ const spinner = require('ora')({});
 const fs = require('fs');
 const path = require('path');
 
-const figma = require('@figma-export/core');
+const figmaExport = require('@figma-export/core');
 
 const resolveNameOrPath = (nameOrPath) => {
     const absolutePath = path.resolve(nameOrPath);
@@ -42,7 +42,7 @@ class ComponentsCommand extends Command {
 
         spinner.start();
 
-        return figma.exportComponents(fileId, {
+        return figmaExport.components(fileId, {
             token: process.env.FIGMA_TOKEN,
             onlyFromPages: page,
             transformers: requirePackages(transformer),
